Move v1 routes back into FastAPI
